+++ uui/src/lib/index.ts
@@ -4,6 +4,7 @@ import Chip from './Chip/chip.svelte';
 import Step from './Stepper/Step.svelte';
 import List from './List/List.svelte';
 import Paper from './Paper/Paper.svelte';
+import Drawer from './Drawer/Drawer.svelte';
 import Theme from './Theme/theme.svelte';
 import Button from './Button/Button.svelte';
 import Dialog from './Dialog/Dialog.svelte';
@@ -31,6 +32,7 @@ export {
   Step,
   List,
   Paper,
+  Drawer,
   Theme,
   Button,
   Dialog,
@@ -130,3 +132,4 @@ export type {
   BreakpointValue
 } from './shared/theme/breakpoints/breakpoints.types';
 export type { PaperProps } from './Paper/Paper.types';
+export type { DrawerProps } from './Drawer/Drawer.types';